chore: fix & improve util types
