+++ packages/blade/src/utils/makeBezier/index.native.ts
@@ -1,0 +1,1 @@
+export * from './makeBezier.native';
