--- packages/blade/src/utils/makeBezier/makeBezier.d.ts
@@ -1,1 +1,0 @@
-export * from './makeBezier.web';
