+++ packages/blade/src/utils/makeTypographySize/index.native.ts
@@ -1,0 +1,1 @@
+export * from './makeTypographySize.native';
