+++ packages/blade/src/utils/makeTypographySize/index.ts
@@ -1,1 +1,1 @@
-export * from './makeTypographySize';
+export * from './makeTypographySize.web';
