--- packages/blade/src/utils/makeTypographySize/makeTypographySize.d.ts
@@ -1,1 +1,0 @@
-export * from './makeTypographySize.web';
