+++ packages/blade/src/utils/testId/index.native.ts
@@ -1,0 +1,1 @@
+export * from './testId.native';
